Separatinf DepthNav files
